move element injection code to the element package
