+++ integration/todo/src/elements/todo-form.element.ts
@@ -1,9 +1,8 @@
-import { inject, injectable } from '@joist/di';
+import { inject } from '@joist/di';
 import { css, html, shadow, listen, tagName, element } from '@joist/element';
 
 import { TodoService, Todo } from '../services/todo.service.js';
 
-@injectable
 @element
 export class TodoFormElement extends HTMLElement {
   @tagName static tagName = 'todo-form';
