+++ packages/di/src/lib.ts
@@ -1,5 +1,4 @@
 export { Injector } from './lib/injector.js';
 export { Provider, ConstructableToken, StaticToken } from './lib/provider.js';
-export { injectable } from './lib/injectable.js';
+export { injectable, INJECTABLE_MAP } from './lib/injectable.js';
 export { inject, Injected } from './lib/inject.js';
-export { defineEnvironment, clearEnvironment } from './lib/environment.js';
